LayoutTests/platform/glib/imported/w3c/web-platform-tests/mathml/relations/css-styling/width-height-005-expected.txt


FAIL mrow assert_greater_than: width is modified expected a number greater than 1 but got 0
FAIL mrow (horizontal overflow) assert_greater_than: width is modified expected a number greater than 1 but got 0
FAIL RTL mrow assert_greater_than: width is modified expected a number greater than 1 but got 0
FAIL RTL mrow (horizontal overflow) assert_greater_than: width is modified expected a number greater than 1 but got 0
FAIL mpadded assert_greater_than: width is modified expected a number greater than 1 but got 0
FAIL mpadded (horizontal overflow) assert_greater_than: width is modified expected a number greater than 1 but got 0
FAIL RTL mpadded assert_greater_than: width is modified expected a number greater than 1 but got 0
FAIL RTL mpadded (horizontal overflow) assert_greater_than: width is modified expected a number greater than 1 but got 0
FAIL msqrt assert_greater_than: width is modified expected a number greater than 1 but got 0
FAIL msqrt (horizontal overflow) assert_greater_than: width is modified expected a number greater than 1 but got 0
FAIL RTL msqrt assert_greater_than: width is modified expected a number greater than 1 but got 0
FAIL RTL msqrt (horizontal overflow) assert_greater_than: width is modified expected a number greater than 1 but got 0
FAIL mroot assert_greater_than: width is modified expected a number greater than 1 but got 0
FAIL mroot (horizontal overflow) assert_greater_than: width is modified expected a number greater than 1 but got 0
FAIL RTL mroot assert_greater_than: width is modified expected a number greater than 1 but got 0
FAIL RTL mroot (horizontal overflow) assert_greater_than: width is modified expected a number greater than 1 but got 0
FAIL msub assert_greater_than: width is modified expected a number greater than 1 but got 0
FAIL msub (horizontal overflow) assert_greater_than: width is modified expected a number greater than 1 but got 0
FAIL RTL msub assert_greater_than: width is modified expected a number greater than 1 but got 0
FAIL RTL msub (horizontal overflow) assert_greater_than: width is modified expected a number greater than 1 but got 0
FAIL msup assert_greater_than: width is modified expected a number greater than 1 but got 0
FAIL msup (horizontal overflow) assert_greater_than: width is modified expected a number greater than 1 but got 0
FAIL RTL msup assert_greater_than: width is modified expected a number greater than 1 but got 0
FAIL RTL msup (horizontal overflow) assert_greater_than: width is modified expected a number greater than 1 but got 0
FAIL msubsup assert_greater_than: width is modified expected a number greater than 1 but got 0
FAIL msubsup (horizontal overflow) assert_greater_than: width is modified expected a number greater than 1 but got 0
FAIL RTL msubsup assert_greater_than: width is modified expected a number greater than 1 but got 0
FAIL RTL msubsup (horizontal overflow) assert_greater_than: width is modified expected a number greater than 1 but got 0
FAIL mmultiscripts assert_greater_than: width is modified expected a number greater than 1 but got 0
FAIL mmultiscripts (horizontal overflow) assert_greater_than: width is modified expected a number greater than 1 but got 0
FAIL RTL mmultiscripts assert_greater_than: width is modified expected a number greater than 1 but got 0
FAIL RTL mmultiscripts (horizontal overflow) assert_greater_than: width is modified expected a number greater than 1 but got 0
PASS mtext
PASS mtext (horizontal overflow)
PASS RTL mtext
PASS RTL mtext (horizontal overflow)
FAIL legacy menclose assert_greater_than: width is modified expected a number greater than 1 but got 0
FAIL legacy menclose (horizontal overflow) assert_greater_than: width is modified expected a number greater than 1 but got 0
FAIL RTL legacy menclose assert_greater_than: width is modified expected a number greater than 1 but got 0
FAIL RTL legacy menclose (horizontal overflow) assert_greater_than: width is modified expected a number greater than 1 but got 0
X
X
X

X
X
X

X
X
X
X
X
X

X
X
X

X
X
X
X
X
X

X
X
X

X
X
X
X
X
X

X
X
X

X
X
X
X
X
X

X
X
X

X
X
X
X
X
X

X
X
X

X
X
X
X
X

X
X

X
X
X
X

X
X

X
X
X
X

X
X

X
X
X
X

X
X

X
X
X
X

X
X

X
X
X
X

X
X

X
X
X
X
X

X
X
X

X
X
X
X
X
X

X
X
X

X
X
X
X
X
X
X
X

X
X
X
X
X

X
X
X
X
X
X
X
X
X
X

X
X
X
X
X

X
X
X
X
X
X X X

X X X

X X X
X X X

X X X

X X X
X
X
X

X
X
X

X
X
X
X
X
X

X
X
X

X
X
X
